user can only delete their own posts

+++ src/main/java/com/makersacademy/acebook/controller/PostsController.java
@@ -12,12 +12,14 @@ import com.makersacademy.acebook.repository.LikeRepository;
 import com.makersacademy.acebook.repository.PostRepository;
 import com.makersacademy.acebook.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.oauth2.core.oidc.user.DefaultOidcUser;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 import org.springframework.web.servlet.view.RedirectView;
 
 import java.util.List;
@@ -88,8 +90,12 @@ public class PostsController {
     }
 
     @DeleteMapping("/posts/{id}")
-    public RedirectView deletePost(@PathVariable Long id, @RequestParam String redirect) {
-        repository.deleteById(id);
+    public RedirectView deletePost(@PathVariable Long id, @RequestParam String redirect, @AuthenticationPrincipal DefaultOidcUser oidcUser) {
+        Optional<User> user = userRepository.findUserByUsername(oidcUser.getEmail());
+        Optional<Post> post = repository.findById(id);
+        if (Objects.equals(user.get().getId(), post.get().getUser().getId())) {
+            repository.deleteById(id);
+        }
         return new RedirectView(redirect);
     }
